+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added https://github.com/Masterminds/sprig functions to templates #1030
 - Add multiple output formats (yaml & json) support. #447
 - More aliases for many wwctl commands
+- Add support to render template using `host` or `$(uname -n)` as the value of `overlay show --render`. #623
 
 ### Changed